+++ api/src/main/java/com/staxpayments/api/repository/InvoiceRepository.kt
@@ -1,13 +1,13 @@
 package com.staxpayments.api.repository
 
 import com.staxpayments.api.models.Invoice
-import com.staxpayments.api.requests.CreateInvoiceBody
+import com.staxpayments.api.requests.InvoicePostRequest
 
 interface InvoiceRepository {
 
-    suspend fun getInvoice(invoiceId: String): Invoice?
+    suspend fun getInvoice(invoiceId: String): Invoice
 
-    suspend fun createInvoice(body: CreateInvoiceBody): Invoice
+    suspend fun createInvoice(body: InvoicePostRequest): Invoice
 
-    suspend fun updateInvoice(invoiceId: String, body: CreateInvoiceBody): Invoice?
+    suspend fun updateInvoice(invoiceId: String, body: InvoicePostRequest): Invoice
 }
